changement type masse et coordonne meteore

+++ OutilMeteores/src/ca/qc/cgmatane/informatique/outilmeteore/Meteore.java
@@ -2,15 +2,15 @@ package ca.qc.cgmatane.informatique.outilmeteore;
 
 public class Meteore 
 {
-	private int[] coordonnees;
+	private float[] coordonnees;
 	private int id;
 	private String description;
-	private int masse;
+	private float masse;
 	private String nom = "";
 	private int annee;
 	
 	
-	public Meteore(int[] coordonnees, int id, int masse, String nom,int annee)
+	public Meteore(float[] coordonnees, int id, float masse, String nom,int annee)
 	{
 		this.coordonnees = coordonnees;
 		this.id = id;
@@ -21,10 +21,10 @@ public class Meteore
 		
 	}
 	
-	public int[] getCoordonnees() {
+	public float[] getCoordonnees() {
 		return coordonnees;
 	}
-	public void setCoordonnees(int[] coordonnees) {
+	public void setCoordonnees(float[] coordonnees) {
 		this.coordonnees = coordonnees;
 	}
 	public int getId() {
@@ -40,11 +40,11 @@ public class Meteore
 		this.description = description;
 	}
 
-	public int getMasse() {
+	public float getMasse() {
 		return masse;
 	}
 
-	public void setMasse(int masse) {
+	public void setMasse(float masse) {
 		this.masse = masse;
 	}
